11 functions fully functional with appropriate error handling

+++ Automate2.py
@@ -89,13 +89,17 @@ def createBucket(vol, bucket):  # done
 # Create a new key inside a bucket
 def createKey(vol, bucket, key):  # done
     command_formation = f"ozone fs -touch o3fs://{bucket}.{vol}.ozone1/{key}"
-    execute(command_formation)
+    err = execute(command_formation)[2]
+    return err.read().decode().strip()
 
 
 # Create n keys at a time inside a bucket
 def createManyKeys(vol, bucket, key_names):  # done
     for key in key_names:
-        createKey(vol, bucket, key)
+        err_catch = createKey(vol, bucket, key)
+        if err_catch != "":
+            return err_catch
+    return ""
 
 
 # Delete a particular volume
@@ -108,13 +112,15 @@ def deleteVolume(vol):  # done
 # Delete a particular bucket from a volume
 def deleteBucket(vol, bucket):  # done
     command_formation = f"ozone sh bucket delete {vol}/{bucket}"
-    execute(command_formation)
+    err = execute(command_formation)[2]
+    return err.read().decode().strip()
 
 
 # Delete a key from the bucket
 def deleteKey(vol, bucket, key_name):  # done
     command_formation = f"ozone sh key delete {vol}/{bucket}/{key_name}"
-    execute(command_formation)
+    err = execute(command_formation)[2]
+    return err.read().decode().strip()
 
 
 # Delete all the keys from the bucket
@@ -167,12 +173,25 @@ def main():
             case 3:
                 showVolumes()
                 vol_name = input("Choose the volume: ")
-                showBuckets(vol_name)
-                bucket_name = input("Choose the bucket: ")
-                key_name = input("Enter the key name: ")
-                createKey(vol_name, bucket_name, key_name)
-                print("Key created successfully.")
-                showKeys(vol_name, bucket_name)
+                err = showBuckets(vol_name)
+                if err != "":
+                    # print(type(err))
+                    print(err)
+                else:
+                    bucket_name = input("Choose the bucket: ")
+                    key_name = input("Enter the key name: ")
+                    err = createKey(vol_name, bucket_name, key_name)
+                    if err == "":
+                        print("Key created successfully.")
+                        showKeys(vol_name, bucket_name)
+                    else:
+                        # print(type(err))
+                        # The error output will have warning lines as well as the actual error message
+                        # Last line is the actual error message, it is in the form " touch: <error message> "
+                        # so, splitlines()[-1] is for getting the last line and partition is for extracting the
+                        # real error message without the "touch"
+                        print(err.splitlines()[-1].partition(": ")[2])
+
             case 4:
                 showVolumes()
             case 5:
@@ -191,28 +210,48 @@ def main():
                 else:
                     print(err)
             case 7:
-                num = int(input("Enter the number of keys(N) you want to add: "))
-                key_names = []
-                for i in range(num):
-                    key_name = input(f"Enter the name of {i + 1} key: ")
-                    key_names.append(key_name)
                 showVolumes()
                 vol_name = input("Choose the volume: ")
-                showBuckets(vol_name)
-                bucket_name = input("Choose the bucket: ")
-                print("Putting keys..Please wait..")
-                createManyKeys(vol_name, bucket_name, key_names)
-                print("All the keys created successfully!")
-                showKeys(vol_name, bucket_name)
+                err = showBuckets(vol_name)
+                if err != "":
+                    # print(type(err))
+                    print(err)
+                else:
+                    bucket_name = input("Choose the bucket: ")
+                    num = int(input("Enter the number of keys(N) you want to add: "))
+                    key_names = []
+                    for i in range(num):
+                        key_name = input(f"Enter the name of {i + 1} key: ")
+                        key_names.append(key_name)
+
+                    print("Putting keys..Please wait..")
+                    err = createManyKeys(vol_name, bucket_name, key_names)
+                    if err == "":
+                        print("All the keys created successfully!")
+                        showKeys(vol_name, bucket_name)
+                    else:
+                        # print(type(err))
+                        # The error output will have warning lines as well as the actual error message
+                        # Last line is the actual error message, it is in the form " touch: <error message> "
+                        # so, splitlines()[-1] is for getting the last line and partition is for extracting the
+                        # real error message without the "touch"
+                        print(err.splitlines()[-1].partition(": ")[2])
+
             case 8:
                 showVolumes()
                 vol_name = input("Choose the volume: ")
-                showBuckets(vol_name)
-                bucket_name = input("Choose the bucket: ")
-                keys = showKeys(vol_name, bucket_name)
-                print("Deleting all keys.. Please wait")
-                deleteAllKeys(vol_name, bucket_name, keys)
-                print("All keys deleted from the specified bucket")
+                err = showBuckets(vol_name)
+                if err == "":
+                    bucket_name = input("Choose the bucket: ")
+                    keys, err = showKeys(vol_name, bucket_name)
+                    if err == "":
+                        print("Deleting all keys.. Please wait")
+                        deleteAllKeys(vol_name, bucket_name, keys)
+                        print("All keys deleted from the specified bucket")
+                    else:
+                        print(err)
+                else:
+                    print(err)
             case 9:
                 showVolumes()
                 vol_name = input("Enter the Volume to delete: ")
@@ -226,21 +265,37 @@ def main():
             case 10:
                 showVolumes()
                 vol_name = input("Choose the volume: ")
-                showBuckets(vol_name)
-                bucket_name = input("Enter the bucket to delete: ")
-                deleteBucket(vol_name, bucket_name)
-                print("Bucket deleted successfully!")
-                showBuckets(vol_name)
+                err = showBuckets(vol_name)
+                if err == "":
+                    bucket_name = input("Enter the bucket to delete: ")
+                    err = deleteBucket(vol_name, bucket_name)
+                    if err == "":
+                        print("Bucket deleted successfully!")
+                        showBuckets(vol_name)
+                    else:
+                        print(err)
+                else:
+                    print(err)
             case 11:
                 showVolumes()
                 vol_name = input("Choose the volume: ")
-                showBuckets(vol_name)
-                bucket_name = input("Choose the bucket: ")
-                showKeys(vol_name, bucket_name)
-                key_name = input("Enter the key name to delete: ")
-                deleteKey(vol_name, bucket_name, key_name)
-                print("Key deleted successfully!")
-                showKeys(vol_name, bucket_name)
+                err = showBuckets(vol_name)
+                if err != "":
+                    # print(type(err))
+                    print(err)
+                else:
+                    bucket_name = input("Choose the bucket: ")
+                    key_list, err = showKeys(vol_name, bucket_name)
+                    if err == "":
+                        key_name = input("Enter the key name to delete: ")
+                        err = deleteKey(vol_name, bucket_name, key_name)
+                        if err == "":
+                            print("Key deleted successfully!")
+                            showKeys(vol_name, bucket_name)
+                        else:
+                            print(err)
+                    else:
+                        print(err)
 
 
 if __name__ == "__main__":
